Create gatekeeper_certification.md
Dear W3C participants,
We are excited to share with you a Gatekeeper (as defined in SPARROW) certification process proposal. These Gatekeepers are necessary trusted servers for online advertising to work without negatively affecting user experience whilst protecting their privacy.
We propose to define together, with all interested parties, a code of conduct that would cover both principles (e.g. no fingerprinting) and rules of operation (e.g. following reporting proposal to disclose information about the ads served).
This code would then be used by certification authorities to certify and audit Gatekeepers to ensure compliance.
We consider this proposal as an actionable first step for the creation of Gatekeepers. This proposal is meant to be challenged. We are open to any suggestion or modification to this proposal to allow trusted entities to be certified.
Please join us in defining this code of conduct and the certification process for the emergence of a trusted server, that would allow for better privacy, user experience, and web monetization.